LayoutTests/imported/w3c/web-platform-tests/html/semantics/embedded-content/the-iframe-element/cross_origin_parentage-expected.txt

Blocked access to external URL http://www1.web-platform.test:8000/html/semantics/embedded-content/the-iframe-element/cross_origin_child.html


FAIL Check the frame heriarchy assert_equals: expected 200 but got 500
FAIL Check the frame heriarchy 1 assert_equals: expected 200 but got 500
FAIL Check the frame heriarchy 2 assert_equals: expected 200 but got 500
